+++ packages/orq-rc/src/orq_ai_sdk/models/listagenttasksop.py
@@ -1227,6 +1227,7 @@ class ListAgentTasksStateTypedDict(TypedDict):
     error: NotRequired[str]
     iteration: NotRequired[int]
     accumulated_execution_time: NotRequired[float]
+    variables: NotRequired[Dict[str, Any]]
 
 
 class ListAgentTasksState(BaseModel):
@@ -1239,6 +1240,8 @@ class ListAgentTasksState(BaseModel):
     iteration: Optional[int] = 1
 
     accumulated_execution_time: Optional[float] = 0
+
+    variables: Optional[Dict[str, Any]] = None
 
 
 class ListAgentTasksMetricsTypedDict(TypedDict):