+++ packages/orq-rc/src/orq_ai_sdk/models/retrievecontactop.py
@@ -103,7 +103,7 @@ class RetrieveContactResponseBody(BaseModel):
     created: Optional[datetime] = None
     r"""The date and time the resource was created"""
 
-    updated: Optional[datetime] = parse_datetime("2025-11-07T12:12:06.586Z")
+    updated: Optional[datetime] = parse_datetime("2025-11-08T09:16:35.289Z")
     r"""The date and time the resource was last updated"""
 
     @model_serializer(mode="wrap")