+++ packages/orq-rc/src/orq_ai_sdk/models/retrievedatasourceop.py
@@ -90,7 +90,7 @@ class RetrieveDatasourceResponseBody(BaseModel):
     r"""The number of chunks in the datasource"""
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "01K9F3PP1MXKPJ1280F0S5JW7X"
+        "01K9HC1ZDYKYTNPDKXX1X8KK9P"
     )
     r"""The unique identifier of the data source"""