+++ packages/orq-rc/src/orq_ai_sdk/models/updatedatasourceop.py
@@ -104,7 +104,7 @@ class UpdateDatasourceResponseBody(BaseModel):
     r"""The number of chunks in the datasource"""
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "01K9F3PP1R2GB1Y2KH075JSQCX"
+        "01K9HC1ZE366GCTGANE1BKBCWA"
     )
     r"""The unique identifier of the data source"""