+++ packages/orq-rc/RELEASES.md
@@ -3909,3 +3909,13 @@ Based on:
 - [python v4.0.0-rc.16] packages/orq-rc
 ### Releases
 - [PyPI v4.0.0-rc.16] https://pypi.org/project/orq-ai-sdk/4.0.0-rc.16 - packages/orq-rc
+
+## 2025-11-08 09:18:16
+### Changes
+Based on:
+- OpenAPI Doc  
+- Speakeasy CLI 1.651.0 (2.745.2) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [python v4.0.0-rc.17] packages/orq-rc
+### Releases
+- [PyPI v4.0.0-rc.17] https://pypi.org/project/orq-ai-sdk/4.0.0-rc.17 - packages/orq-rc